deployment/security-agent/backend/tests/requirements.txt
boto3==1.34.162
PyJWT==2.8.0